Disable Counter metric's associated _created gauge
The Counter type metrics will emit `_total` and `_created` time series,
where the `_created` timeseries is a contant of gague type.

With the introduction of OpenMetrics, the `_created` timeseries will be
converted to `_created_total` and it is a bit messy.

Their purpose related to detecting if the `_total` counter was reset,
but there is detection logic for that without a `_created` timestamp as
well.

So I propose we drop it to make things simpler. Having done that, an
opentelemetry-collector's served metrics will be the same as a
prometheus-client's served metrics.